+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/factoryIrUtil.kt
@@ -9,6 +9,7 @@ import dev.zacsweers.metro.compiler.ir.IrAnnotation
 import dev.zacsweers.metro.compiler.ir.IrMetroContext
 import dev.zacsweers.metro.compiler.ir.IrTypeKey
 import dev.zacsweers.metro.compiler.ir.annotationClass
+import dev.zacsweers.metro.compiler.ir.annotationsIn
 import dev.zacsweers.metro.compiler.ir.copyParameterDefaultValues
 import dev.zacsweers.metro.compiler.ir.createIrBuilder
 import dev.zacsweers.metro.compiler.ir.deepRemapperFor
@@ -430,8 +431,15 @@ internal fun IrFunction.addParameters(
           defaultValue = context.createIrBuilder(symbol).run { irExprBody(stubExpression()) }
         }
       }
-      .applyIf(copyQualifiers) {
-        param.typeKey.qualifier?.let { annotations += it.ir.deepCopyWithSymbols() }
+      .apply {
+        // Propagate @OptionalBinding if present
+        param.asValueParameter
+          .annotationsIn(context.metroSymbols.classIds.optionalBindingAnnotations)
+          .firstOrNull()
+          ?.let { annotations += it.deepCopyWithSymbols() }
+        if (copyQualifiers) {
+          param.typeKey.qualifier?.let { annotations += it.ir.deepCopyWithSymbols() }
+        }
       }
       .also { onParam(param.typeKey, it) }
   }