+++ CHANGELOG.md
@@ -46,6 +46,7 @@ annotation class ViewModelKey(val value: KClass<out ViewModel> = Nothing::class)
 - **[IR]** Fix a severity conversion compat function call for Kotlin 2.3.20+.
 - **[IR]** Ensure stable sort of output `SuspiciousUnusedMultibinding` locations.
 - **[IR]** Don't skip dynamic keys inherited from parent graphs when working with dynamic graphs.
+- **[IR]** Propagate `@OptionalBinding` annotations to generated static factory creators if present.
 - **[Runtime]** `IntoSet` and `IntoMap` no longer have a `Target` of `AnnotationTarget.CLASS`
 
 ### Changes
@@ -62,9 +63,12 @@ annotation class ViewModelKey(val value: KClass<out ViewModel> = Nothing::class)
 
 Special thanks to the following contributors for contributing to this release!
 
+- [@Asapha](https://github.com/Asapha)
 - [@ChristianKatzmann](https://github.com/ChristianKatzmann)
 - [@grandstaish](https://github.com/grandstaish)
 - [@jonamireh](https://github.com/jonamireh)
+- [@kevinguitar](https://github.com/kevinguitar)
+- [@svenjacobs](https://github.com/svenjacobs)
 
 0.11.4
 ------